experiments/computer-use-cua/pyproject.toml
[project]
name = "computer-use-cua"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cua-agent[all]>=0.2.1",
    "cua-computer>=0.1.24",
    "openai>=1.76.2",
    "python-dotenv>=1.1.0",
]
